updated fetchData function
the fetchData function in actions considers the dataType that was
requested.
for SectionBounds, TileBounds, MatchesWithinGroup, and
MatchesOutsideGroup, the data from all of the layers in the range
requested is fetched using Promise.all and then put into a dictionary
that maps the Z layer to its data.
for all other dataType requests, the json that is returned upon
fetching is that is returned.