added componentDidUpdate
when the component updates (which is after each render), a check is
made to see if tileData was generated and the visualization was
rendered.
if tileData is present but the visualization was not rendered, events
listeners are added, the three.js scene is created
(generateVisualization()) and rendered is set to true so that the scene
will not be recreated.